requirements.txt.txt
# Core LLM + LangChain
langchain==0.1.14
langchain-community==0.0.27
langchain-core==0.1.40
huggingface_hub==0.22.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# PDF & Markdown handling
PyMuPDF==1.23.6           # To read PDFs
python-docx==1.1.0        # To parse .docx policy files
markdown==3.4.1           # For Markdown conversion
weasyprint==61.2          # To generate PDFs from Markdown

# Streamlit frontend
streamlit==1.33.0

# Utilities
python-dotenv==1.0.1      # For .env loading
tqdm==4.66.2              # For progress bars if needed
pandas==2.2.2             # For CSV-based incident data
scikit-learn==1.3.2       # For any ML logic (if future upgrade)

# Web dependencies (for Streamlit + PDF gen)
cairosvg==2.7.1           # For PDF rendering
lxml==5.1.0               # Required by weasyprint
html5lib==1.1             # HTML parsing for PDFs

# Optional (good for debugging + future growth)
openai==1.30.1            # In case fallback to OpenAI LLMs
